+++ plugins/@sanity/embeddings-index-ui/src/embeddingsIndexDashboard/hooks.ts
@@ -1,0 +1,36 @@
+import {useMemo} from 'react'
+import type {ObjectSchemaType, Schema} from 'sanity'
+
+import {isType} from '../utils/types'
+
+const defaultProjection = '{...}'
+
+export function useDefaultIndex(schema: Schema, dataset: string) {
+  const defaultFilter = useMemo(
+    () =>
+      `_type in [${schema
+        .getTypeNames()
+        .map((n) => schema.get(n))
+        .filter((schemaType): schemaType is ObjectSchemaType =>
+          Boolean(schemaType && isType(schemaType, 'document')),
+        )
+        .filter(
+          (documentType) =>
+            !documentType.name.startsWith('sanity.') &&
+            !documentType.name.startsWith('assist.') &&
+            documentType.name !== 'document',
+        )
+        .map((documentType) => `"${documentType.name}"`)
+        .join(',\n')}]`,
+    [schema],
+  )
+
+  return useMemo(
+    () => ({
+      dataset,
+      projection: defaultProjection,
+      filter: defaultFilter,
+    }),
+    [defaultFilter, dataset],
+  )
+}