+++ plugins/sanity-naive-html-serializer/src/BaseDocumentSerializer/index.ts
@@ -1,8 +1,8 @@
-import {PortableTextTypeComponent, toHTML} from '@portabletext/to-html'
-import {SanityDocument, TypedObject, Schema} from 'sanity'
+import {type PortableTextTypeComponent, toHTML} from '@portabletext/to-html'
+import type {SanityDocument, TypedObject, Schema} from 'sanity'
 
 import {defaultStopTypes, customSerializers} from '../BaseSerializationConfig'
-import {TranslationLevel, SerializerClosure} from '../types'
+import type {TranslationLevel, SerializerClosure} from '../types'
 import {fieldFilter, internationalizedArrayFilter, languageObjectFieldFilter} from './fieldFilters'
 
 const META_FIELDS = ['_key', '_type', '_id', '_weak']