+++ plugins/sanity-plugin-mux-input/src/components/PageSelector.tsx
@@ -1,6 +1,6 @@
 import {ChevronLeftIcon, ChevronRightIcon} from '@sanity/icons'
 import {Button, Label} from '@sanity/ui'
-import {Dispatch, SetStateAction, useEffect} from 'react'
+import {type Dispatch, type SetStateAction, useEffect} from 'react'
 
 const PageSelector = (props: {
   page: number